+++ src/javascript/esutil.ts
@@ -100,7 +100,7 @@ export function closureType(
       code: 'no-closure-type',
       message:
           `Unable to determine closure type for expression of type ${
-                                                                     node.type
+    node.type
                                                                    }`,
       severity: Severity.WARNING, sourceRange
     });