+++ src/test/static/analysis/class/classes.js
@@ -1,15 +1,24 @@
 class Element extends HTMLElement { }
 customElements.define('my-element', Element);
 
-/** @polymerElement */
+/**
+ * @customElement
+ * @polymer
+ */
 class AnnotatedElement { static get is() {return 'annotated-elem'} }
 
-/** @polymerMixin */
+/**
+ * @mixinFunction
+ * @polymer
+ */
 function Mixin(superClass) {
   return class Mixin extends SuperClass { }
 }
 
-/** @polymerMixin */
+/**
+ * @mixinFunction
+ * @polymer
+ */
 AnnotatedMixin = CreateMixinSomehow();