+++ src/test/static/polymer2-old-jsdoc/test-element-11.js
@@ -1,0 +1,14 @@
+
+class MyElement extends Polymer.Element {
+  static get is() { return 'my-app'; }
+  static get properties() {
+    return {
+      prop1: {
+        type: String,
+        value: 'my-app'
+      }
+    };
+  }
+}
+
+window.customElements.define(MyElement.is, MyElement);