+++ src/test/static/polymer2-old-jsdoc/test-element-3.js
@@ -1,0 +1,32 @@
+/**
+ * @polymerElement
+ * @extends Polymer.Element
+ */
+class BaseElement extends Polymer.Element {
+  static get properties() {
+    return {
+      foo: {
+        notify: true,
+        type: String,
+      },
+    };
+  }
+}
+
+/**
+ * @polymerElement
+ * @extends BaseElement
+ */
+class SubElement extends BaseElement {
+  static get is() {
+    return 'sub-element';
+  }
+  static get properties() {
+    return {
+      bar: {
+        notify: true,
+        type: String,
+      },
+    };
+  }
+}