+++ src/test/static/polymer2-old-jsdoc/test-element-6.js
@@ -1,0 +1,12 @@
+/**
+ * @polymerElement
+ * @extends Polymer.Element
+ * @mixes Mixin2
+ * @mixes Mixin1
+ */
+class TestElement extends Mixin1
+(Mixin2(Polymer.Element)) {
+  static get is() {
+    return 'test-element';
+  }
+}