+++ src/test/static/polymer2-old-jsdoc/test-mixin-3.js
@@ -1,0 +1,16 @@
+/**
+ * @polymerMixin
+ * @memberof Polymer
+ */
+const TestMixin = function(superclass) {
+  return class extends superclass {
+    static get properties() {
+      return {
+        foo: {
+          notify: true,
+          type: String,
+        },
+      };
+    }
+  }
+}