+++ src/test/static/polymer2-old-jsdoc/test-mixin-4.js
@@ -1,0 +1,18 @@
+/**
+ * @polymerMixin
+ * @memberof Polymer
+ */
+let TestMixin;
+
+function Foo(superclass) {
+  return class extends superclass {
+    static get properties() {
+      return {
+        foo: {
+          notify: true,
+          type: String,
+        },
+      };
+    }
+  }
+}