+++ src/test/static/polymer2-old-jsdoc/test-mixin-6.js
@@ -1,0 +1,6 @@
+/**
+ * @polymerMixin
+ * @memberof Polymer
+ */
+function TestMixin() {
+}