+++ src/test/static/polymer2-old-jsdoc/test-mixin-8.js
@@ -1,0 +1,14 @@
+/**
+ * @polymerMixin
+ * @memberof Polymer
+ */
+Polymer.TestMixin = (superclass) => class extends superclass {
+  static get properties() {
+    return {
+      foo: {
+        notify: true,
+        type: String,
+      }
+    };
+  }
+}