+++ src/test/static/polymer2/test-element-5.js
@@ -1,5 +1,6 @@
 /**
- * @polymerElement
+ * @customElement
+ * @polymer
  */
 class TestElement extends Polymer.Element {
   static get observedAttributes() {