+++ src/test/static/polymer2/test-mixin-3.js
@@ -1,5 +1,6 @@
 /**
- * @polymerMixin
+ * @polymer
+ * @mixinFunction
  * @memberof Polymer
  */
 const TestMixin = function(superclass) {