+++ src/test/static/polymer2/test-mixin-8.js
@@ -1,5 +1,6 @@
 /**
- * @polymerMixin
+ * @polymer
+ * @mixinFunction
  * @memberof Polymer
  */
 Polymer.TestMixin = (superclass) => class extends superclass {